Replace Alamofire with direct URLSession usage
#### Sunday
Alamofire is very helpful in writing HTTP requests with predermined formats and content types for body and response data; its not so good at allowing requests to adapt to body and response content types at runtime. Another issue is it’s error handling generally provides little to no data from the actual response.

Sunday was written to handle these types of scenarios and other shortcomings of Alamofire when working with complex and dynamic REST APIs and thus ended up using very little of the Alamofire framework in general.

This commit replaces all use of Alamofire with direct use of the URLSession framework and/or homegrown replacements.

Changes:
* `RequestManager` now uses RxSwift types exclusively
* `URLSession` is extended with RxSwift methods for requests, responses & streams
  * Required to have its delegate be a `SessionDelegate` when using Sunday extensions
* `NetworkRequestManager` is now implemented using `URLSession`s RxSwift extensions
* `EventSource` is now implemented using `URLSession`s RxSwift extensions
* `URLEncoder` implements `MediaTypeEncoder`  for `application/x-www-form-urlencoded`
* `ServiceLocator` is a service browser implementation for easily finding network services

#### SudayServer
SundayServer was not using Alamofire directly but it haas been enhanced to support more HTTP verbs and provide easier debugging information.

Changes:
* Numerous objects (e.g. HTTP.Request) are now string covertible for debugging
* `HTTPServer` now better supports services
* `HEAD` method can now be handled in `Routable`s